+++ goodgame.gg/src/main/java/fourjo/idle/goodgame/gg/web/dto/AdminUserSearchDTO.java
@@ -1,0 +1,23 @@
+package fourjo.idle.goodgame.gg.web.dto;
+
+import lombok.Data;
+
+@Data
+public class AdminUserSearchDTO {
+
+int page;
+
+    private String searchValue;
+
+    private String limit;
+
+    private int count;
+
+    private int index;
+
+    public void setIndex() {
+        index = (page-1) * count;
+    }
+
+
+}
