+++ goodgame.gg/src/main/resources/mappers/admin.xml
@@ -31,7 +31,7 @@
 
     </select>
 
-    <select id="AdminUserSearchList"  parameterType="string" resultType="fourjo.idle.goodgame.gg.web.dto.UserDto">
+    <select id="AdminUserSearch"  parameterType="fourjo.idle.goodgame.gg.web.dto.AdminUserSearchDTO" resultType="fourjo.idle.goodgame.gg.web.dto.UserDto">
     select
     user_index as userIndex,
     user_id as userId,
@@ -42,7 +42,18 @@
     user_email as userEmail,
     role_id as roleId
     from user_mst
-    where user_nick like CONCAT('%', #{user_nick}, '%')
+        <if test='searchValue != null and searchValue !=""'>
+            where (
+            user_id like CONCAT('%', #{searchValue}, '%')
+            or user_nick like CONCAT('%', #{searchValue}, '%')
+            or user_email like CONCAT('%', #{searchValue}, '%')
+            )
+        </if>
+        order by
+        user_index desc
+        <if test='limit == "Y"'>
+            limit #{index}, #{count}
+        </if>
 
 </select>
